Move ParseRevocationList from parser.go to crl.go
